Convert cube to be android aware
