+++ examples/cube/Cargo.toml
@@ -7,6 +7,9 @@ edition = "2021"
 publish = false
 rust-version = "1.57"
 
+[lib]
+crate-type = ["lib", "cdylib"]
+
 [[bin]]
 name = "cube"
 path = "src/main.rs"
@@ -25,7 +28,7 @@ rend3-routine = { version = "^0.2.0", path = "../../rend3-routine" }
 # Framework that deals with the event loop, setting up the renderer, and platform differences.
 rend3-framework = { version = "^0.2.0", path = "../../rend3-framework" }
 # windowing
-winit = { version = "0.25", features = ["web-sys"] }
+winit = { version = "0.26" }
 
 [target.'cfg(target_arch = "wasm32")'.dependencies]
 console_log = "0.2"
@@ -34,3 +37,14 @@ js-sys = "0.3"
 web-sys = "0.3"
 wasm-bindgen = "=0.2.78"
 wasm-bindgen-futures = "0.4"
+
+[target.'cfg(target_os = "android")'.dependencies]
+ndk-glue = { version = "0.5", features = ["logger"] }
+
+[package.metadata.andoird]
+build_targets = ["aarch64-linux-android"]
+
+[package.metadata.android.sdk]
+min_sdk_version = 28
+target_sdk_version = 28
+compile_sdk_version = 28
