+++ examples/cube/src/lib.rs
@@ -1,0 +1,186 @@
+use std::sync::Arc;
+
+fn vertex(pos: [f32; 3]) -> glam::Vec3 {
+    glam::Vec3::from(pos)
+}
+
+fn create_mesh() -> rend3::types::Mesh {
+    let vertex_positions = [
+        // far side (0.0, 0.0, 1.0)
+        vertex([-1.0, -1.0, 1.0]),
+        vertex([1.0, -1.0, 1.0]),
+        vertex([1.0, 1.0, 1.0]),
+        vertex([-1.0, 1.0, 1.0]),
+        // near side (0.0, 0.0, -1.0)
+        vertex([-1.0, 1.0, -1.0]),
+        vertex([1.0, 1.0, -1.0]),
+        vertex([1.0, -1.0, -1.0]),
+        vertex([-1.0, -1.0, -1.0]),
+        // right side (1.0, 0.0, 0.0)
+        vertex([1.0, -1.0, -1.0]),
+        vertex([1.0, 1.0, -1.0]),
+        vertex([1.0, 1.0, 1.0]),
+        vertex([1.0, -1.0, 1.0]),
+        // left side (-1.0, 0.0, 0.0)
+        vertex([-1.0, -1.0, 1.0]),
+        vertex([-1.0, 1.0, 1.0]),
+        vertex([-1.0, 1.0, -1.0]),
+        vertex([-1.0, -1.0, -1.0]),
+        // top (0.0, 1.0, 0.0)
+        vertex([1.0, 1.0, -1.0]),
+        vertex([-1.0, 1.0, -1.0]),
+        vertex([-1.0, 1.0, 1.0]),
+        vertex([1.0, 1.0, 1.0]),
+        // bottom (0.0, -1.0, 0.0)
+        vertex([1.0, -1.0, 1.0]),
+        vertex([-1.0, -1.0, 1.0]),
+        vertex([-1.0, -1.0, -1.0]),
+        vertex([1.0, -1.0, -1.0]),
+    ];
+
+    let index_data: &[u32] = &[
+        0, 1, 2, 2, 3, 0, // far
+        4, 5, 6, 6, 7, 4, // near
+        8, 9, 10, 10, 11, 8, // right
+        12, 13, 14, 14, 15, 12, // left
+        16, 17, 18, 18, 19, 16, // top
+        20, 21, 22, 22, 23, 20, // bottom
+    ];
+
+    rend3::types::MeshBuilder::new(vertex_positions.to_vec())
+        .with_indices(index_data.to_vec())
+        .build()
+        .unwrap()
+}
+
+#[derive(Default)]
+struct CubeExample {
+    object_handle: Option<rend3::types::ObjectHandle>,
+    directional_light_handle: Option<rend3::types::DirectionalLightHandle>,
+}
+
+impl rend3_framework::App for CubeExample {
+    const DEFAULT_SAMPLE_COUNT: rend3::types::SampleCount = rend3::types::SampleCount::One;
+
+    fn setup(
+        &mut self,
+        _window: &winit::window::Window,
+        renderer: &Arc<rend3::Renderer>,
+        _routines: &Arc<rend3_framework::DefaultRoutines>,
+        _surface: &Arc<rend3::types::Surface>,
+        _surface_format: rend3::types::TextureFormat,
+    ) {
+        // Create mesh and calculate smooth normals based on vertices
+        let mesh = create_mesh();
+
+        // Add mesh to renderer's world.
+        //
+        // All handles are refcounted, so we only need to hang onto the handle until we make an object.
+        let mesh_handle = renderer.add_mesh(mesh);
+
+        // Add PBR material with all defaults except a single color.
+        let material = rend3_routine::material::PbrMaterial {
+            albedo: rend3_routine::material::AlbedoComponent::Value(glam::Vec4::new(0.0, 0.5, 0.5, 1.0)),
+            ..rend3_routine::material::PbrMaterial::default()
+        };
+        let material_handle = renderer.add_material(material);
+
+        // Combine the mesh and the material with a location to give an object.
+        let object = rend3::types::Object {
+            mesh: mesh_handle,
+            material: material_handle,
+            transform: glam::Mat4::IDENTITY,
+        };
+        // Creating an object will hold onto both the mesh and the material
+        // even if they are deleted.
+        //
+        // We need to keep the object handle alive.
+        self.object_handle = Some(renderer.add_object(object));
+
+        let view_location = glam::Vec3::new(3.0, 3.0, -5.0);
+        let view = glam::Mat4::from_euler(glam::EulerRot::XYZ, -0.55, 0.5, 0.0);
+        let view = view * glam::Mat4::from_translation(-view_location);
+
+        // Set camera's location
+        renderer.set_camera_data(rend3::types::Camera {
+            projection: rend3::types::CameraProjection::Perspective { vfov: 60.0, near: 0.1 },
+            view,
+        });
+
+        // Create a single directional light
+        //
+        // We need to keep the directional light handle alive.
+        self.directional_light_handle = Some(renderer.add_directional_light(rend3::types::DirectionalLight {
+            color: glam::Vec3::ONE,
+            intensity: 10.0,
+            // Direction will be normalized
+            direction: glam::Vec3::new(-1.0, -4.0, 2.0),
+            distance: 400.0,
+        }));
+    }
+
+    fn handle_event(
+        &mut self,
+        window: &winit::window::Window,
+        renderer: &Arc<rend3::Renderer>,
+        routines: &Arc<rend3_framework::DefaultRoutines>,
+        surface: &Arc<rend3::types::Surface>,
+        event: rend3_framework::Event<'_, ()>,
+        control_flow: impl FnOnce(winit::event_loop::ControlFlow),
+    ) {
+        match event {
+            // Close button was clicked, we should close.
+            rend3_framework::Event::WindowEvent {
+                event: winit::event::WindowEvent::CloseRequested,
+                ..
+            } => {
+                control_flow(winit::event_loop::ControlFlow::Exit);
+            }
+            rend3_framework::Event::MainEventsCleared => {
+                window.request_redraw();
+            }
+            // Render!
+            rend3_framework::Event::RedrawRequested(_) => {
+                // Get a frame
+                let frame = rend3::util::output::OutputFrame::Surface {
+                    surface: Arc::clone(surface),
+                };
+                // Ready up the renderer
+                let (cmd_bufs, ready) = renderer.ready();
+
+                // Lock the routines
+                let pbr_routine = rend3_framework::lock(&routines.pbr);
+                let tonemapping_routine = rend3_framework::lock(&routines.tonemapping);
+
+                // Build a rendergraph
+                let mut graph = rend3::RenderGraph::new();
+
+                // Add the default rendergraph without a skybox
+                rend3_routine::add_default_rendergraph(
+                    &mut graph,
+                    &ready,
+                    &pbr_routine,
+                    None,
+                    &tonemapping_routine,
+                    Self::DEFAULT_SAMPLE_COUNT,
+                );
+
+                // Dispatch a render using the built up rendergraph!
+                graph.execute(renderer, frame, cmd_bufs, &ready);
+            }
+            // Other events we don't care about
+            _ => {}
+        }
+    }
+}
+
+#[cfg_attr(target_os = "android", ndk_glue::main(backtrace = "on", logger(level = "debug")))]
+pub fn main() {
+    let app = CubeExample::default();
+    rend3_framework::start(
+        app,
+        winit::window::WindowBuilder::new()
+            .with_title("cube-example")
+            .with_maximized(true),
+    );
+}
